requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
logging
pymongo[srv]==3.12
##-e .